HF_DEPLOY/requirements.txt
gradio
torch
torchvision
scikit-learn
huggingface-hub
pandas
pillow
numpy
